Reduced named message hash from ulong to uint
Truncated named message hash down to 4B. Deprecated support for 8B RPC hashes as a side effect.

+++ com.unity.netcode.gameobjects/Runtime/Messaging/Messages/NamedMessage.cs
@@ -2,7 +2,7 @@ namespace Unity.Netcode
 {
     internal struct NamedMessage : INetworkMessage
     {
-        public ulong Hash;
+        public uint Hash;
         public FastBufferWriter SendData;
 
         private FastBufferReader m_ReceiveData;
